refactor: Split up useNodeLayout drag specific functionality
